refactor(vscode): lift diff overlay out of behavior.ts
Slice of #767 — extracts the per-card diff overlay (Side / Overlay /
Onion modes plus client-side pixel diff) from `behavior.ts` into a
typed `diffOverlay.ts` module.

Same DOM shape, same CSS classes, same async pixel-walk; the host
hands in a `DiffOverlayConfig` so the module doesn't have to reach
into the closed-over `state.diffMode` / `vscode.setState` plumbing.
The three call shapes (`payload`, `errorMessage`, both null = loading
sentinel) match the previous control flow exactly.

`behavior.ts` shrinks from 3208 to 2936 lines (-272 LOC). No
behaviour change.

Refs #767.
